Extract rx-based workflow functionality to separate module
