+++ libraries/rib-base/src/main/java/com/badoo/ribs/core/Node.kt
@@ -5,7 +5,6 @@ import android.os.Parcelable
 import android.util.SparseArray
 import androidx.annotation.CallSuper
 import androidx.annotation.MainThread
-import androidx.annotation.VisibleForTesting
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.LifecycleRegistry
@@ -30,10 +29,6 @@ import com.badoo.ribs.core.plugin.ViewAware
 import com.badoo.ribs.core.plugin.ViewLifecycleAware
 import com.badoo.ribs.core.view.RibView
 import com.badoo.ribs.util.RIBs
-import com.jakewharton.rxrelay2.BehaviorRelay
-import com.jakewharton.rxrelay2.PublishRelay
-import io.reactivex.Observable
-import io.reactivex.Single
 import java.util.concurrent.CopyOnWriteArrayList
 
 /**
@@ -88,12 +83,9 @@ open class Node<V : RibView>(
 
     private val savedInstanceState = buildParams.savedInstanceState?.getBundle(BUNDLE_KEY)
     internal val externalLifecycleRegistry = LifecycleRegistry(this)
-    val detachSignal = BehaviorRelay.create<Unit>()
 
     val tag: String = this::class.java.name
     val children = CopyOnWriteArrayList<Node<*>>()
-    private val childrenAttachesRelay: PublishRelay<Node<*>> = PublishRelay.create()
-    val childrenAttaches: Observable<Node<*>> = childrenAttachesRelay.hide()
 
     internal open val lifecycleManager = LifecycleManager(this)
 
@@ -201,7 +193,6 @@ open class Node<V : RibView>(
             detachChildNode(child)
         }
 
-        detachSignal.accept(Unit)
         isPendingDetach = false
     }
 
@@ -216,8 +207,11 @@ open class Node<V : RibView>(
         children.add(child)
         lifecycleManager.onAttachChild(child)
         child.onAttach()
-        childrenAttachesRelay.accept(child)
+        onAttachChildNode(child)
         plugins.filterIsInstance<SubtreeChangeAware>().forEach { it.onAttachChild(child) }
+    }
+
+    open fun onAttachChildNode(child: Node<*>) {
     }
 
     private fun verifyNotRoot(child: Node<*>) {
@@ -381,79 +375,4 @@ open class Node<V : RibView>(
 
         return null
     }
-
-    /**
-     * Executes an action and remains on the same hierarchical level
-     *
-     * @return the current workflow element
-     */
-    protected inline fun <reified T> executeWorkflow(
-        crossinline action: () -> Unit
-    ): Single<T> = Single.fromCallable {
-            action()
-            this as T
-        }
-        .takeUntil(detachSignal.firstOrError())
-
-    @VisibleForTesting
-    internal inline fun <reified T> executeWorkflowInternal(
-        crossinline action: () -> Unit
-    ) : Single<T> = executeWorkflow(action)
-
-    /**
-     * Executes an action and transitions to another workflow element
-     *
-     * @param action an action that's supposed to result in the attach of a child (e.g. router.push())
-     *
-     * @return the child as the expected workflow element, or error if expected child was not found
-     */
-    @SuppressWarnings("LongMethod")
-    protected inline fun <reified T> attachWorkflow(
-        crossinline action: () -> Unit
-    ): Single<T> = Single.fromCallable {
-            action()
-            val childNodesOfExpectedType = children.filterIsInstance<T>()
-            if (childNodesOfExpectedType.isEmpty()) {
-                Single.error<T>(
-                    IllegalStateException(
-                        "Expected child of type [${T::class.java}] was not found after executing action. " +
-                            "Check that your action actually results in the expected child. " +
-                            "Child count: ${children.size}. " +
-                            "Last child is: [${children.lastOrNull()}]. " +
-                            "All children: $children"
-                    )
-                )
-            } else {
-                Single.just(childNodesOfExpectedType.last())
-            }
-        }
-        .flatMap { it }
-        .takeUntil(detachSignal.firstOrError())
-
-    @VisibleForTesting
-    internal inline fun <reified T> attachWorkflowInternal(
-        crossinline action: () -> Unit
-    ) : Single<T> = attachWorkflow(action)
-
-    /**
-     * Waits until a certain child is attached and returns it as the expected workflow element, or
-     * returns it immediately if it's already available.
-     *
-     * @return the child as the expected workflow element
-     */
-    protected inline fun <reified T> waitForChildAttached(): Single<T> =
-        Single.fromCallable {
-            val childNodesOfExpectedType = children.filterIsInstance<T>()
-            if (childNodesOfExpectedType.isEmpty()) {
-                childrenAttaches.ofType(T::class.java).firstOrError()
-            } else {
-                Single.just(childNodesOfExpectedType.last())
-            }
-        }
-        .flatMap { it }
-        .takeUntil(detachSignal.firstOrError())
-
-    @VisibleForTesting
-    internal inline fun <reified T> waitForChildAttachedInternal() : Single<T> =
-        waitForChildAttached()
 }
