+++ libraries/rib-portal/src/main/java/com/badoo/ribs/portal/PortalNode.kt
@@ -1,6 +1,5 @@
 package com.badoo.ribs.portal
 
-import com.badoo.ribs.core.Node
 import com.badoo.ribs.core.Rib
 import com.badoo.ribs.core.modality.AncestryInfo
 import com.badoo.ribs.core.modality.BuildParams
@@ -9,13 +8,14 @@ import com.badoo.ribs.portal.PortalRouter.Configuration
 import com.badoo.ribs.portal.PortalRouter.Configuration.Content
 import com.badoo.ribs.routing.source.backstack.BackStackFeature
 import com.badoo.ribs.routing.source.backstack.operation.push
+import com.badoo.ribs.workflows.rx.RxWorkflowNode
 import io.reactivex.Single
 
 internal class PortalNode(
     buildParams: BuildParams<*>,
     plugins: List<Plugin>,
     private val backStack: BackStackFeature<Configuration>
-) : Node<Nothing>(
+) : RxWorkflowNode<Nothing>(
     buildParams = buildParams,
     viewFactory = null,
     plugins = plugins
