+++ libraries/rib-workflows-rx/build.gradle
@@ -7,28 +7,12 @@ androidExtensions {
     experimental = true
 }
 
-tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
-    kotlinOptions {
-        freeCompilerArgs += "-Xopt-in=com.badoo.ribs.annotation.ExperimentalApi"
-    }
-}
-
 dependencies {
-    api deps.android.appCompat
-
-    implementation deps.apt.javaxInject
-
-    implementation deps.android.recyclerView
-    implementation deps.android.lifecycleCommon
-    compileOnly deps.android.annotations
     implementation deps.external.kotlinStdlib
-    implementation deps.external.mviCore
-    implementation deps.external.mviCoreAndroid
-    implementation deps.external.reactiveStreams
     implementation deps.external.rxrelay2
     implementation deps.external.rxjava2
+
     implementation project(":libraries:rib-base")
-    implementation project(":libraries:rib-workflows-rx")
 
     testImplementation deps.android.annotations
     testImplementation deps.test.junit
