+++ sandbox/build.gradle
@@ -67,6 +67,7 @@ dependencies {
     implementation project(":libraries:rib-base")
     implementation project(":libraries:rib-portal")
     implementation project(":libraries:rib-recyclerview")
+    implementation project(":libraries:rib-workflows-rx")
 
     if (useCompose) {
         implementation project(":libraries:rib-compose")
