+++ sandbox/src/main/java/com/badoo/ribs/sandbox/rib/hello_world/HelloWorldNode.kt
@@ -3,12 +3,12 @@ package com.badoo.ribs.sandbox.rib.hello_world
 import android.util.Log
 import com.badoo.ribs.clienthelper.connector.Connectable
 import com.badoo.ribs.clienthelper.connector.NodeConnector
-import com.badoo.ribs.core.Node
 import com.badoo.ribs.core.modality.BuildParams
 import com.badoo.ribs.core.plugin.Plugin
 import com.badoo.ribs.core.view.RibView
 import com.badoo.ribs.sandbox.rib.hello_world.HelloWorld.Input
 import com.badoo.ribs.sandbox.rib.hello_world.HelloWorld.Output
+import com.badoo.ribs.workflows.rx.RxWorkflowNode
 import io.reactivex.Single
 
 class HelloWorldNode(
@@ -16,7 +16,7 @@ class HelloWorldNode(
     viewFactory: ((RibView) -> HelloWorldView?)?,
     plugins: List<Plugin> = emptyList(),
     connector: NodeConnector<Input, Output> = NodeConnector()
-) : Node<HelloWorldView>(
+) : RxWorkflowNode<HelloWorldView>(
     buildParams = buildParams,
     viewFactory = viewFactory,
     plugins = plugins
