+++ settings.gradle
@@ -15,6 +15,7 @@ include ':libraries:rib-base-test-activity'
 include ':libraries:rib-debug-utils'
 include ':libraries:rib-portal'
 include ':libraries:rib-recyclerview'
+include ':libraries:rib-workflows-rx'
 include ':sandbox'
 include ':templates:release-0.8'
 include ':templates:release-0.12'
